src/main/java/com/denso/demo/service/SkateBoardServiceImpl.java
package com.denso.demo.service;

import com.denso.demo.domain.SkateBoard;
import com.denso.demo.entity.SkateBoardEntity;
import com.denso.demo.repository.SkateBoardRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.Collections;
import java.util.List;

@Service
@RequiredArgsConstructor
public class SkateBoardServiceImpl implements SkateBoardService {

    private final SkateBoardRepository skateBoardRepository;

    @Override
    public List<SkateBoard> getAllAvailableSkateBoards() {
        List<SkateBoardEntity> skateBoardEntityList = skateBoardRepository.findAll();

        if (skateBoardEntityList.isEmpty()) {
            return Collections.emptyList();
        }
        return null;
    }
}
